Sql_alchemy/requirements.txt
mysql-connector-python==8.1.0
pandas==2.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
